pyproject.toml
# =============================================================================
# pyproject.toml - FastAPI Boilerplate Configuration
#
# CONFIGURATION: Update project name, description, and author information.
# =============================================================================

[project]
name = "fastapi-boilerplate"
version = "1.0.0"
description = "FastAPI boilerplate with PostgreSQL + MongoDB support"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
keywords = ["fastapi", "api", "boilerplate", "postgresql", "mongodb"]

dependencies = [
    # FastAPI & Web Framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "python-multipart>=0.0.9",
    "httpx>=0.27.0",

    # Database - PostgreSQL
    # NOTE: Remove if not using PostgreSQL
    "sqlalchemy[asyncio]>=2.0.35",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",

    # Database - MongoDB
    # NOTE: Remove if not using MongoDB
    "motor>=3.6.0",
    "beanie>=1.26.0",
    "pymongo>=4.6.1",

    # Authentication & Security
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",

    # Validation & Utilities
    "email-validator>=2.2.0",
    "python-dotenv>=1.0.0",

    # Monitoring & Logging
    "structlog>=24.4.0",
]

[dependency-groups]
dev = [
    # Testing
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "pytest-mock>=3.14.0",
    "faker>=30.0.0",
    "factory-boy>=3.3.0",

    # Code Quality
    "ruff>=0.7.0",
    "mypy>=1.13.0",
    "types-passlib>=1.7.7",
    "types-python-jose>=3.3.4",

    # Dev Tools
    "ipython>=8.20.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

# =============================================================================
# Ruff Configuration (Linting & Formatting)
# =============================================================================
[tool.ruff]
line-length = 100
target-version = "py312"
src = ["app", "tests"]
exclude = [
    ".git",
    ".venv",
    "__pycache__",
    "build",
    "dist",
    "*.egg-info",
    "alembic/versions",
]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long
    "B008",  # function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"tests/**/*.py" = ["ARG"]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

# =============================================================================
# MyPy Configuration (Type Checking)
# =============================================================================
[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_generics = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true

[[tool.mypy.overrides]]
module = [
    "motor.*",
    "beanie.*",
    "pymongo.*",
]
ignore_missing_imports = true

# =============================================================================
# Pytest Configuration (Testing)
# =============================================================================
[tool.pytest.ini_options]
minversion = "8.0"
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--cov=app",
    "--cov-report=term-missing",
    "--cov-report=html",
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "slow: Slow running tests",
]

# =============================================================================
# Coverage Configuration
# =============================================================================
[tool.coverage.run]
source = ["app"]
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    "*/site-packages/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@(abc\\.)?abstractmethod",
]
